extract get-flattened-dependencies logic to a new file
